requirements.txt
Pillow==9.3.0
watermark.py==0.0.3.post3
